+++ src/Sharpen.Engine/SharpenSuggestions/CSharp70/OutVariables/DiscardOutVariablesInObjectCreations.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.CodeAnalysis.CSharp.Syntax;
+
+namespace Sharpen.Engine.SharpenSuggestions.CSharp70.OutVariables
+{
+    internal sealed class DiscardOutVariablesInObjectCreations : BaseUseOutVariables<ObjectCreationExpressionSyntax>
+    {
+        private DiscardOutVariablesInObjectCreations() :base(true) { }
+
+        public static readonly DiscardOutVariablesInObjectCreations Instance = new DiscardOutVariablesInObjectCreations();
+    }
+}